Made it run more than twice as fast!
The observation is that when a hailstone number is even and N is the number of trailing zero-bits, then performing the next N steps is equivalent to simply right shifting by N bits. Using a lookup table to give me the number of trailing zero bits (modulo 64) and shifting by that amount is what gives the speed increase.

The lookup table is 64 single byte entries, on the assumptions that (a) a cache line is 64 bytes wide, so this will fit on it; and (b) that the cost of unpacking the single byte entries into a size_t will be less than the cost of accessing a table which is spread over sizeof(size_t) cache lines.

I haven't done any tests to validate these assumptions yet; nor have I made any effort to ensure that the start of the table lies on a 64 byte boundary. So there's still room for improvement!