Twice as fast again!
Added a table to record previously seen sequence lengths and use them to short-circuit calculation of newer sequence lengths.

I picked a table size of 1,048,576 elements (each of which is a size_t) pretty much arbitrarily. There's some potential gains to be had by tuning the size of this table.

Another thing to investigate is only storing sequence lengths for odd numbers, since even numbers reduce down to odd numbers very quickly already. This would effectively double the number of sequences we can store for quick lookup later at the cost of a single right-shift per lookup - which may or may not be worth it.