Another big speed improvement - for SOME cases...
We don't need to calculate lengths for even numbers the long way.
Instead, iterate over the odd numbers calculating their length and then
for each one calculate the length for every power of two multiple of it
by double it and incrementing the length until we pass the upper bound.

We need to be a bit careful at the bottom of the range, to make sure we
don't accidentally skip some numbers. So instead, we process every
number - odd or even - between lower and 2*lower. Above that we only
process odd numbers.

If the range we're calculating falls entirely into [lower, 2*lower) then
the conditional checks for the doubling steps are unnecessary and slow
things down, so there's a separate code path for that case.

It might be worth splitting the input into 3 ranges instead of two:
[lower, 2*lower), [2*lower, upper/2) and [upper/2, upper]
That would mean we could skip the conditional checks in the last range
as well. But that's for later...