Store length counts, not bucket counts, while processing.
The maximum length we'll need to store for numbers up to 2^32 is less than 1132, which is tiny enough that we can store an individual count for each length without worrying about the space. That means we avoid calculating a couple of things in our inner loops: the bucket number (a subtraction & a division) as well as the if-statement to check whether we're over the maximum length or not.

The bucket totals are now calculated later, after we have all the length counts.
 
Net result: 10% speed increase, give or take a bit. Sweet!