Replaced the trailing zero bits LUT with a compiler intrinsic.
GCC has an intrinsic function i__builtin_ctzll, for counting the number of
trailing zero bits in an unsigned long long. Using this means we don't need to
calculate the lookup table. We can also avoid checking whether the number is
odd inside the length calculation loop - we can guarantee it is because we
always get all of the trailing zeros in one go, rather than at most the last 8.

Overall the code runs about 1.4 times as fast with this change.